Fix Paths
The repo hosting the package is called `metals-sublime`
but in package control the package is called `LSP-metals`.
So we have to fix the paths to keep things working.

Also update the readme for people wanting the install the package manually (via git)

See https://github.com/sublimelsp/LSP/issues/1074#issuecomment-650395484